requirements.txt
alembic==1.4.3
asgiref==3.2.3
atari-py==0.2.6
blinker==1.4
boto==2.49.0
boto3==1.10.47
botocore==1.13.47
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==6.7
cloudpickle==1.6.0
ConfigArgParse==1.4
cycler==0.10.0
Django==3.0.2
docutils==0.15.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
gensim==3.8.1
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
gym==0.17.3
h5py==2.9.0
idna==2.8
itsdangerous==0.24
jieba==0.40
Jinja2==2.11.3
jmespath==0.9.4
kiwisolver==1.0.1
locust==1.4.3
Mako==1.0.6
MarkupSafe==1.0
matplotlib==3.0.2
mkl-fft==1.0.12
mkl-random==1.0.2
msgpack==1.0.2
mysql==0.0.2
mysqlclient==2.0.3
nltk==3.4.5
numpy==1.16.4
nvidia-ml-py3==7.352.0
olefile==0.44
opencv-python==4.4.0.42
pandas==0.25.3
pdfminer.six==20200104
Pillow==6.1.0
psutil==5.8.0
pycparser==2.20
pycryptodome==3.9.4
pyglet==1.5.0
pyparsing==2.3.0
PyPDF2==1.26.0
pyreadline==2.1
python-dateutil==2.7.5
python-editor==1.0.4
pytz==2019.3
pywin32==300
pyzmq==22.0.3
requests==2.22.0
s3transfer==0.2.1
scipy==1.4.1
six==1.12.0
smart-open==1.9.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.22
sqlparse==0.3.0
statsmodels==0.9.0
torch==1.5.1
torchvision==0.6.1
urllib3==1.25.7
Werkzeug==1.0.1
wincertstore==0.2
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.2.0
